pyproject.toml
[project]
name = "agent-builder-demo"
version = "0.1.0"
description = "Agent Builder Demo - 基于 LangChain DeepAgents 的智能 Agent 构建工具"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=1.0.2",
    "langchain-anthropic>=1.0.3",
    "langchain_tavily>=0.2.13",
    "pydantic>=2.0.0",
    "tavily-python>=0.5.0",
    "httpx>=0.28.1",
    "markdownify>=1.2.0",
    "deepagents>=0.2.6",
    "python-dotenv>=1.0.0",
    "streamlit>=1.30.0"
]
